Switch Windows default back to OpenGL, change handling of the option
The trick with creating a dummy file for the OpenGL vs. ANGLE setting
was somewhat attractive, because we didn't have to load/parse the main
config file a second time. The problem is that SettingsComponent is a
singleton, and it needs to be created at a specific point in the
program. We can't just create it earlier.

But it created an awkward requirement to restart PMP twice when editing
the config file. It also makes it much harder to change the default
setting.

Abandon the previous approach, and let it load the main config file
twice. The logic for this is in readPreinitValue(). (Not bothering with
the case if we need this for something else - loading the config file is
cheap.) This duplicates some config file traversal logic, but
fortunately it's only 1 line of code.